Move out of auth
- Moved change password and update profile
