+++ app/Http/Controllers/Account.php
@@ -3,9 +3,11 @@ declare(strict_types=1);
 
 namespace App\Http\Controllers;
 
+use App\Actions\Account\ChangePassword;
 use App\Actions\Account\DeleteAccount;
 use App\Actions\Account\DeleteBudgetAccount;
 use App\Actions\Account\Reset;
+use App\Actions\Account\UpdateProfile;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -16,6 +18,49 @@ use Illuminate\Support\Facades\Auth;
  */
 class Account extends Controller
 {
+    public function changePassword(Request $request)
+    {
+        $this->bootstrap($request);
+
+        $user = $this->api->getAuthUser();
+
+        if ($user['status'] !== 200) {
+            abort(404, 'Unable to fetch your account information from the API');
+        }
+
+        return view(
+            'account.change-password',
+            [
+                'user' => $user
+            ]
+        );
+    }
+
+    public function changePasswordProcess(Request $request)
+    {
+        $this->bootstrap($request);
+
+        $action = new ChangePassword();
+        $result = $action(
+            $this->api,
+            $request->only(['password', 'password_confirmation'])
+        );
+
+        if ($result === 204) {
+            return redirect()->route('account.index')
+                ->with('status', 'password-changed');
+        }
+
+        if ($result === 422) {
+            return redirect()
+                ->route('account.change-password')
+                ->withInput()
+                ->with('validation.errors', $action->getValidationErrors());
+        }
+
+        abort($result, $action->getMessage());
+    }
+
     public function index(Request $request)
     {
         $this->bootstrap($request);
@@ -158,5 +203,48 @@ class Account extends Controller
         Auth::guard()->logout();
 
         return redirect()->route('landing');
+    }
+
+    public function updateProfile(Request $request)
+    {
+        $this->bootstrap($request);
+
+        $user = $this->api->getAuthUser();
+
+        if ($user['status'] !== 200) {
+            abort(404, 'Unable to fetch your account information from the API');
+        }
+
+        return view(
+            'account.update-profile',
+            [
+                'user' => $user
+            ]
+        );
+    }
+
+    public function updateProfileProcess(Request $request)
+    {
+        $this->bootstrap($request);
+
+        $action = new UpdateProfile();
+        $result = $action(
+            $this->api,
+            $request->only(['name', 'email'])
+        );
+
+        if ($result === 204) {
+            return redirect()->route('account.index')
+                ->with('status', 'profile-updated');
+        }
+
+        if ($result === 422) {
+            return redirect()
+                ->route('account.update-profile')
+                ->withInput()
+                ->with('validation.errors', $action->getValidationErrors());
+        }
+
+        abort($result, $action->getMessage());
     }
 }
